@@ -21,9 +21,11 @@
     "@babel/runtime": "^7.22.3",
     "babel-loader": "^9.1.2",
     "concurrently": "^8.1.0",
+    "css-loader": "^6.8.1",
     "html-webpack-plugin": "^5.5.2",
     "nodemon": "^2.0.22",
-    "webpack": "^5.85.0",
+    "style-loader": "^3.3.3",
+    "webpack": "^5.86.0",
     "webpack-cli": "^5.1.1",
     "webpack-dev-server": "^4.15.0"
   },
